refactor(sections): relocate section routes
