+++ apps/ui/src/components/pages/application/utils/useSectionForm.ts
@@ -17,8 +17,8 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { SectionRoutesValues } from '@/pages/AppRouter';
 import { useApplicationContext } from '@/providers/context/application/ApplicationContext';
+import { SectionRoutesValues } from '@pcgl-daco/validation';
 import { Form } from 'antd';
 import { useEffect } from 'react';
