+++ apps/ui/src/pages/AppRouter.tsx
@@ -36,21 +36,7 @@ import DashboardPage from '@/pages/dashboard';
 import HomePage from '@/pages/index';
 import ManageApplicationsPage from '@/pages/manage/applications';
 import { ApplicationContextProvider } from '@/providers/context/application/ApplicationContextProvider';
-
-export const SectionRoutes = {
-	INTRO: 'intro',
-	APPLICANT: 'applicant',
-	INSTITUTIONAL: 'institutional',
-	COLLABORATORS: 'collaborators',
-	PROJECT: 'project',
-	STUDY: 'study',
-	ETHICS: 'ethics',
-	AGREEMENT: 'agreement',
-	APPENDICES: 'appendices',
-	SIGN: 'sign',
-} as const;
-
-export type SectionRoutesValues = (typeof SectionRoutes)[keyof typeof SectionRoutes];
+import { SectionRoutes, SectionRoutesValues } from '@pcgl-daco/validation';
 
 export interface ApplicationSectionRouteTypes {
 	route: SectionRoutesValues;
